Add models with observable objects to network availability and offline service.
